+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - selfTest not restoring range and not clearing FIFO.
 - FIFO reading algorithm.
 - clearFifo possible infinite loop.
+- getAccelerationScaleFactor not returning NAN for an unexpected range value.
 
 ## [1.3.0] - 2025-08-21
 ### Added